Refactor mo_form_utilities.js
